Add files via upload
The objective of designing this robot is simply to facilitate the humans in the future for security purposes. The main objective of the project was to command and control the Robot wirelessly by the GUI Application. The main task of this project was basically divided in two parts: 
(1)to program the arduino microcontroller mounted on Robot interfaced which accepts input from the cell phone connected to another cell phone farther away and enables us to control the motion and other specifications of the robot. 
(2)to program the GUI Application in android platform which would enable us to connect to robotic interface.